WIP: intptr_t for syscalls

+++ code/cgame/cg_syscalls.c
@@ -36,7 +36,7 @@ Q_EXPORT void dllEntry( intptr_t (QDECL  *syscallptr)( intptr_t arg,... ) ) {
 }
 
 
-int PASSFLOAT( float x ) {
+intptr_t PASSFLOAT( float x ) {
 	floatint_t fi;
 	fi.f = x;
 	return fi.i;
@@ -58,7 +58,7 @@ int		trap_Milliseconds( void ) {
 }
 
 void	trap_Cvar_Register( vmCvar_t *vmCvar, const char *varName, const char *defaultValue, int flags ) {
-	syscall( CG_CVAR_REGISTER, vmCvar, varName, defaultValue, flags );
+	syscall( CG_CVAR_REGISTER, vmCvar, varName, defaultValue, (intptr_t) flags );
 }
 
 void	trap_Cvar_Update( vmCvar_t *vmCvar ) {
@@ -70,7 +70,7 @@ void	trap_Cvar_Set( const char *var_name, const char *value ) {
 }
 
 void trap_Cvar_VariableStringBuffer( const char *var_name, char *buffer, int bufsize ) {
-	syscall( CG_CVAR_VARIABLESTRINGBUFFER, var_name, buffer, bufsize );
+	syscall( CG_CVAR_VARIABLESTRINGBUFFER, var_name, buffer, (intptr_t) bufsize );
 }
 
 int		trap_Argc( void ) {
@@ -78,31 +78,31 @@ int		trap_Argc( void ) {
 }
 
 void	trap_Argv( int n, char *buffer, int bufferLength ) {
-	syscall( CG_ARGV, n, buffer, bufferLength );
+	syscall( CG_ARGV, n, buffer, (intptr_t) bufferLength );
 }
 
 void	trap_Args( char *buffer, int bufferLength ) {
-	syscall( CG_ARGS, buffer, bufferLength );
+	syscall( CG_ARGS, buffer, (intptr_t) bufferLength );
 }
 
 int		trap_FS_FOpenFile( const char *qpath, fileHandle_t *f, fsMode_t mode ) {
-	return syscall( CG_FS_FOPENFILE, qpath, f, mode );
+	return syscall( CG_FS_FOPENFILE, qpath, f, (intptr_t) mode );
 }
 
 void	trap_FS_Read( void *buffer, int len, fileHandle_t f ) {
-	syscall( CG_FS_READ, buffer, len, f );
+	syscall( CG_FS_READ, buffer, (intptr_t) len, (intptr_t) f );
 }
 
 void	trap_FS_Write( const void *buffer, int len, fileHandle_t f ) {
-	syscall( CG_FS_WRITE, buffer, len, f );
+	syscall( CG_FS_WRITE, buffer, (intptr_t) len, (intptr_t) f );
 }
 
 void	trap_FS_FCloseFile( fileHandle_t f ) {
-	syscall( CG_FS_FCLOSEFILE, f );
+	syscall( CG_FS_FCLOSEFILE, (intptr_t) f );
 }
 
 int trap_FS_Seek( fileHandle_t f, long offset, int origin ) {
-	return syscall( CG_FS_SEEK, f, offset, origin );
+	return syscall( CG_FS_SEEK, (intptr_t) f, (intptr_t) offset, (intptr_t) origin );
 }
 
 void	trap_SendConsoleCommand( const char *text ) {
@@ -134,7 +134,7 @@ int		trap_CM_NumInlineModels( void ) {
 }
 
 clipHandle_t trap_CM_InlineModel( int index ) {
-	return syscall( CG_CM_INLINEMODEL, index );
+	return syscall( CG_CM_INLINEMODEL, (intptr_t) index );
 }
 
 clipHandle_t trap_CM_TempBoxModel( const vec3_t mins, const vec3_t maxs ) {
@@ -146,80 +146,80 @@ clipHandle_t trap_CM_TempCapsuleModel( const vec3_t mins, const vec3_t maxs ) {
 }
 
 int		trap_CM_PointContents( const vec3_t p, clipHandle_t model ) {
-	return syscall( CG_CM_POINTCONTENTS, p, model );
+	return syscall( CG_CM_POINTCONTENTS, p, (intptr_t) model );
 }
 
 int		trap_CM_TransformedPointContents( const vec3_t p, clipHandle_t model, const vec3_t origin, const vec3_t angles ) {
-	return syscall( CG_CM_TRANSFORMEDPOINTCONTENTS, p, model, origin, angles );
+	return syscall( CG_CM_TRANSFORMEDPOINTCONTENTS, p, (intptr_t) model, origin, angles );
 }
 
 void	trap_CM_BoxTrace( trace_t *results, const vec3_t start, const vec3_t end,
 						  const vec3_t mins, const vec3_t maxs,
 						  clipHandle_t model, int brushmask ) {
-	syscall( CG_CM_BOXTRACE, results, start, end, mins, maxs, model, brushmask );
+	syscall( CG_CM_BOXTRACE, results, start, end, mins, maxs, (intptr_t) model, (intptr_t) brushmask );
 }
 
 void	trap_CM_CapsuleTrace( trace_t *results, const vec3_t start, const vec3_t end,
 						  const vec3_t mins, const vec3_t maxs,
 						  clipHandle_t model, int brushmask ) {
-	syscall( CG_CM_CAPSULETRACE, results, start, end, mins, maxs, model, brushmask );
+	syscall( CG_CM_CAPSULETRACE, results, start, end, mins, maxs, model, (intptr_t) brushmask );
 }
 
 void	trap_CM_TransformedBoxTrace( trace_t *results, const vec3_t start, const vec3_t end,
 						  const vec3_t mins, const vec3_t maxs,
 						  clipHandle_t model, int brushmask,
 						  const vec3_t origin, const vec3_t angles ) {
-	syscall( CG_CM_TRANSFORMEDBOXTRACE, results, start, end, mins, maxs, model, brushmask, origin, angles );
+	syscall( CG_CM_TRANSFORMEDBOXTRACE, results, start, end, mins, maxs, (intptr_t) model, (intptr_t) brushmask, origin, angles );
 }
 
 void	trap_CM_TransformedCapsuleTrace( trace_t *results, const vec3_t start, const vec3_t end,
 						  const vec3_t mins, const vec3_t maxs,
 						  clipHandle_t model, int brushmask,
 						  const vec3_t origin, const vec3_t angles ) {
-	syscall( CG_CM_TRANSFORMEDCAPSULETRACE, results, start, end, mins, maxs, model, brushmask, origin, angles );
+	syscall( CG_CM_TRANSFORMEDCAPSULETRACE, results, start, end, mins, maxs, (intptr_t) model, (intptr_t) brushmask, origin, angles );
 }
 
 int		trap_CM_MarkFragments( int numPoints, const vec3_t *points, 
 				const vec3_t projection,
 				int maxPoints, vec3_t pointBuffer,
 				int maxFragments, markFragment_t *fragmentBuffer ) {
-	return syscall( CG_CM_MARKFRAGMENTS, numPoints, points, projection, maxPoints, pointBuffer, maxFragments, fragmentBuffer );
+	return syscall( CG_CM_MARKFRAGMENTS, (intptr_t) numPoints, points, projection, (intptr_t) maxPoints, pointBuffer, (intptr_t) maxFragments, fragmentBuffer );
 }
 
 void	trap_S_StartSound( vec3_t origin, int entityNum, int entchannel, sfxHandle_t sfx ) {
-	syscall( CG_S_STARTSOUND, origin, entityNum, entchannel, sfx );
+	syscall( CG_S_STARTSOUND, origin, (intptr_t) entityNum, (intptr_t) entchannel, (intptr_t) sfx );
 }
 
 void	trap_S_StartLocalSound( sfxHandle_t sfx, int channelNum ) {
-	syscall( CG_S_STARTLOCALSOUND, sfx, channelNum );
+	syscall( CG_S_STARTLOCALSOUND, (intptr_t) sfx, (intptr_t) channelNum );
 }
 
 void	trap_S_ClearLoopingSounds( qboolean killall ) {
-	syscall( CG_S_CLEARLOOPINGSOUNDS, killall );
+	syscall( CG_S_CLEARLOOPINGSOUNDS, (intptr_t) killall );
 }
 
 void	trap_S_AddLoopingSound( int entityNum, const vec3_t origin, const vec3_t velocity, sfxHandle_t sfx ) {
-	syscall( CG_S_ADDLOOPINGSOUND, entityNum, origin, velocity, sfx );
+	syscall( CG_S_ADDLOOPINGSOUND, (intptr_t) entityNum, origin, velocity, (intptr_t) sfx );
 }
 
 void	trap_S_AddRealLoopingSound( int entityNum, const vec3_t origin, const vec3_t velocity, sfxHandle_t sfx ) {
-	syscall( CG_S_ADDREALLOOPINGSOUND, entityNum, origin, velocity, sfx );
+	syscall( CG_S_ADDREALLOOPINGSOUND, (intptr_t) entityNum, origin, velocity, (intptr_t) sfx );
 }
 
 void	trap_S_StopLoopingSound( int entityNum ) {
-	syscall( CG_S_STOPLOOPINGSOUND, entityNum );
+	syscall( CG_S_STOPLOOPINGSOUND, (intptr_t) entityNum );
 }
 
 void	trap_S_UpdateEntityPosition( int entityNum, const vec3_t origin ) {
-	syscall( CG_S_UPDATEENTITYPOSITION, entityNum, origin );
+	syscall( CG_S_UPDATEENTITYPOSITION, (intptr_t) entityNum, origin );
 }
 
 void	trap_S_Respatialize( int entityNum, const vec3_t origin, vec3_t axis[3], int inwater ) {
-	syscall( CG_S_RESPATIALIZE, entityNum, origin, axis, inwater );
+	syscall( CG_S_RESPATIALIZE, (intptr_t) entityNum, origin, axis, (intptr_t) inwater );
 }
 
 sfxHandle_t	trap_S_RegisterSound( const char *sample, qboolean compressed ) {
-	return syscall( CG_S_REGISTERSOUND, sample, compressed );
+	return syscall( CG_S_REGISTERSOUND, sample, (intptr_t) compressed );
 }
 
 void	trap_S_StartBackgroundTrack( const char *intro, const char *loop ) {
@@ -247,7 +247,7 @@ qhandle_t trap_R_RegisterShaderNoMip( const char *name ) {
 }
 
 void trap_R_RegisterFont(const char *fontName, int pointSize, fontInfo_t *font) {
-	syscall(CG_R_REGISTERFONT, fontName, pointSize, font );
+	syscall(CG_R_REGISTERFONT, fontName, (intptr_t) pointSize, font );
 }
 
 void	trap_R_ClearScene( void ) {
@@ -259,11 +259,11 @@ void	trap_R_AddRefEntityToScene( const refEntity_t *re ) {
 }
 
 void	trap_R_AddPolyToScene( qhandle_t hShader , int numVerts, const polyVert_t *verts ) {
-	syscall( CG_R_ADDPOLYTOSCENE, hShader, numVerts, verts );
+	syscall( CG_R_ADDPOLYTOSCENE, (intptr_t) hShader, (intptr_t) numVerts, verts );
 }
 
 void	trap_R_AddPolysToScene( qhandle_t hShader , int numVerts, const polyVert_t *verts, int num ) {
-	syscall( CG_R_ADDPOLYSTOSCENE, hShader, numVerts, verts, num );
+	syscall( CG_R_ADDPOLYSTOSCENE, (intptr_t) hShader, (intptr_t) numVerts, verts, (intptr_t) num );
 }
 
 int		trap_R_LightForPoint( vec3_t point, vec3_t ambientLight, vec3_t directedLight, vec3_t lightDir ) {
@@ -288,16 +288,16 @@ void	trap_R_SetColor( const float *rgba ) {
 
 void	trap_R_DrawStretchPic( float x, float y, float w, float h, 
 							   float s1, float t1, float s2, float t2, qhandle_t hShader ) {
-	syscall( CG_R_DRAWSTRETCHPIC, PASSFLOAT(x), PASSFLOAT(y), PASSFLOAT(w), PASSFLOAT(h), PASSFLOAT(s1), PASSFLOAT(t1), PASSFLOAT(s2), PASSFLOAT(t2), hShader );
+	syscall( CG_R_DRAWSTRETCHPIC, PASSFLOAT(x), PASSFLOAT(y), PASSFLOAT(w), PASSFLOAT(h), PASSFLOAT(s1), PASSFLOAT(t1), PASSFLOAT(s2), PASSFLOAT(t2), (intptr_t) hShader );
 }
 
 void	trap_R_ModelBounds( clipHandle_t model, vec3_t mins, vec3_t maxs ) {
-	syscall( CG_R_MODELBOUNDS, model, mins, maxs );
+	syscall( CG_R_MODELBOUNDS, (intptr_t) model, mins, maxs );
 }
 
 int		trap_R_LerpTag( orientation_t *tag, clipHandle_t mod, int startFrame, int endFrame, 
 					   float frac, const char *tagName ) {
-	return syscall( CG_R_LERPTAG, tag, mod, startFrame, endFrame, PASSFLOAT(frac), tagName );
+	return syscall( CG_R_LERPTAG, tag, (intptr_t) mod, (intptr_t) startFrame, (intptr_t) endFrame, PASSFLOAT(frac), tagName );
 }
 
 void	trap_R_RemapShader( const char *oldShader, const char *newShader, const char *timeOffset ) {
@@ -317,11 +317,11 @@ void		trap_GetCurrentSnapshotNumber( int *snapshotNumber, int *serverTime ) {
 }
 
 qboolean	trap_GetSnapshot( int snapshotNumber, snapshot_t *snapshot ) {
-	return syscall( CG_GETSNAPSHOT, snapshotNumber, snapshot );
+	return syscall( CG_GETSNAPSHOT, (intptr_t) snapshotNumber, snapshot );
 }
 
 qboolean	trap_GetServerCommand( int serverCommandNumber ) {
-	return syscall( CG_GETSERVERCOMMAND, serverCommandNumber );
+	return syscall( CG_GETSERVERCOMMAND, (intptr_t) serverCommandNumber );
 }
 
 int			trap_GetCurrentCmdNumber( void ) {
@@ -329,15 +329,15 @@ int			trap_GetCurrentCmdNumber( void ) {
 }
 
 qboolean	trap_GetUserCmd( int cmdNumber, usercmd_t *ucmd ) {
-	return syscall( CG_GETUSERCMD, cmdNumber, ucmd );
+	return syscall( CG_GETUSERCMD, (intptr_t) cmdNumber, ucmd );
 }
 
 void		trap_SetUserCmdValue( int stateValue, float sensitivityScale ) {
-	syscall( CG_SETUSERCMDVALUE, stateValue, PASSFLOAT(sensitivityScale) );
+	syscall( CG_SETUSERCMDVALUE, (intptr_t) stateValue, PASSFLOAT(sensitivityScale) );
 }
 
 void		testPrintInt( char *string, int i ) {
-	syscall( CG_TESTPRINTINT, string, i );
+	syscall( CG_TESTPRINTINT, string, (intptr_t) i );
 }
 
 void		testPrintFloat( char *string, float f ) {
@@ -349,7 +349,7 @@ int trap_MemoryRemaining( void ) {
 }
 
 qboolean trap_Key_IsDown( int keynum ) {
-	return syscall( CG_KEY_ISDOWN, keynum );
+	return syscall( CG_KEY_ISDOWN, (intptr_t) keynum );
 }
 
 int trap_Key_GetCatcher( void ) {
@@ -357,7 +357,7 @@ int trap_Key_GetCatcher( void ) {
 }
 
 void trap_Key_SetCatcher( int catcher ) {
-	syscall( CG_KEY_SETCATCHER, catcher );
+	syscall( CG_KEY_SETCATCHER, (intptr_t) catcher );
 }
 
 int trap_Key_GetKey( const char *binding ) {
@@ -373,15 +373,15 @@ int trap_PC_LoadSource( const char *filename ) {
 }
 
 int trap_PC_FreeSource( int handle ) {
-	return syscall( CG_PC_FREE_SOURCE, handle );
+	return syscall( CG_PC_FREE_SOURCE, (intptr_t) handle );
 }
 
 int trap_PC_ReadToken( int handle, pc_token_t *pc_token ) {
-	return syscall( CG_PC_READ_TOKEN, handle, pc_token );
+	return syscall( CG_PC_READ_TOKEN, (intptr_t) handle, pc_token );
 }
 
 int trap_PC_SourceFileAndLine( int handle, char *filename, int *line ) {
-	return syscall( CG_PC_SOURCE_FILE_AND_LINE, handle, filename, line );
+	return syscall( CG_PC_SOURCE_FILE_AND_LINE, (intptr_t) handle, filename, line );
 }
 
 void	trap_S_StopBackgroundTrack( void ) {
@@ -398,31 +398,31 @@ void trap_SnapVector( float *v ) {
 
 // this returns a handle.  arg0 is the name in the format "idlogo.roq", set arg1 to NULL, alteredstates to qfalse (do not alter gamestate)
 int trap_CIN_PlayCinematic( const char *arg0, int xpos, int ypos, int width, int height, int bits) {
-  return syscall(CG_CIN_PLAYCINEMATIC, arg0, xpos, ypos, width, height, bits);
+  return syscall(CG_CIN_PLAYCINEMATIC, arg0, (intptr_t) xpos, (intptr_t) ypos, (intptr_t) width, (intptr_t) height, (intptr_t) bits);
 }
  
 // stops playing the cinematic and ends it.  should always return FMV_EOF
 // cinematics must be stopped in reverse order of when they are started
 e_status trap_CIN_StopCinematic(int handle) {
-  return syscall(CG_CIN_STOPCINEMATIC, handle);
+  return syscall(CG_CIN_STOPCINEMATIC, (intptr_t) handle);
 }
 
 
 // will run a frame of the cinematic but will not draw it.  Will return FMV_EOF if the end of the cinematic has been reached.
 e_status trap_CIN_RunCinematic (int handle) {
-  return syscall(CG_CIN_RUNCINEMATIC, handle);
+  return syscall(CG_CIN_RUNCINEMATIC, (intptr_t) handle);
 }
  
 
 // draws the current frame
 void trap_CIN_DrawCinematic (int handle) {
-  syscall(CG_CIN_DRAWCINEMATIC, handle);
+  syscall(CG_CIN_DRAWCINEMATIC, (intptr_t) handle);
 }
  
 
 // allows you to resize the animation dynamically
 void trap_CIN_SetExtents (int handle, int x, int y, int w, int h) {
-  syscall(CG_CIN_SETEXTENTS, handle, x, y, w, h);
+  syscall(CG_CIN_SETEXTENTS, (intptr_t) handle, (intptr_t) x, (intptr_t) y, (intptr_t) w, (intptr_t) h);
 }
 
 /*
@@ -440,7 +440,7 @@ qboolean trap_getCameraInfo( int time, vec3_t *origin, vec3_t *angles) {
 */
 
 qboolean trap_GetEntityToken( char *buffer, int bufferSize ) {
-	return syscall( CG_GET_ENTITY_TOKEN, buffer, bufferSize );
+	return syscall( CG_GET_ENTITY_TOKEN, buffer, (intptr_t) bufferSize );
 }
 
 qboolean trap_R_inPVS( const vec3_t p1, const vec3_t p2 ) {
